third_party/blink/web_tests/http/tests/loading/redirect-methods-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
frame "0" - BeginNavigation request to 'about:blank', http method GET
frame "0" - DidStartNavigation
frame "0" - ReadyToCommitNavigation
frame "0" - didCommitLoadForFrame
frame "0" - didFinishDocumentLoadForFrame
frame "0" - didHandleOnloadEventsForFrame
frame "0" - didFinishLoadForFrame
frame "0" - BeginNavigation request to 'http://127.0.0.1:8000/loading/resources/redirect-methods-form.html', http method GET
frame "0" - DidStartNavigation
frame "0" - ReadyToCommitNavigation
frame "0" - didCommitLoadForFrame
frame "0" - didFinishDocumentLoadForFrame
frame "0" - didHandleOnloadEventsForFrame
frame "0" - didFinishLoadForFrame
frame "0" - BeginNavigation request to 'http://127.0.0.1:8000/loading/resources/redirect-methods-result.php', http method POST
frame "0" - DidStartNavigation
frame "0" - ReadyToCommitNavigation
frame "0" - didCommitLoadForFrame
frame "0" - didFinishDocumentLoadForFrame
frame "1" - BeginNavigation request to 'about:blank', http method GET
frame "1" - DidStartNavigation
frame "1" - ReadyToCommitNavigation
frame "1" - didCommitLoadForFrame
frame "1" - didFinishDocumentLoadForFrame
frame "1" - didHandleOnloadEventsForFrame
frame "1" - didFinishLoadForFrame
frame "1" - BeginNavigation request to 'http://127.0.0.1:8000/loading/resources/redirect-methods-form.html', http method GET
frame "1" - DidStartNavigation
frame "0" - didHandleOnloadEventsForFrame
frame "0" - didFinishLoadForFrame
frame "1" - ReadyToCommitNavigation
frame "1" - didCommitLoadForFrame
frame "1" - didFinishDocumentLoadForFrame
frame "1" - didHandleOnloadEventsForFrame
frame "1" - didFinishLoadForFrame
frame "1" - BeginNavigation request to 'http://127.0.0.1:8000/loading/resources/redirect-methods-result.php', http method POST
frame "1" - DidStartNavigation
frame "1" - ReadyToCommitNavigation
frame "1" - didCommitLoadForFrame
frame "1" - didFinishDocumentLoadForFrame
frame "2" - BeginNavigation request to 'about:blank', http method GET
frame "2" - DidStartNavigation
frame "2" - ReadyToCommitNavigation
frame "2" - didCommitLoadForFrame
frame "2" - didFinishDocumentLoadForFrame
frame "2" - didHandleOnloadEventsForFrame
frame "2" - didFinishLoadForFrame
frame "2" - BeginNavigation request to 'http://127.0.0.1:8000/loading/resources/redirect-methods-form.html', http method GET
frame "2" - DidStartNavigation
frame "1" - didHandleOnloadEventsForFrame
frame "1" - didFinishLoadForFrame
frame "2" - ReadyToCommitNavigation
frame "2" - didCommitLoadForFrame
frame "2" - didFinishDocumentLoadForFrame
frame "2" - didHandleOnloadEventsForFrame
frame "2" - didFinishLoadForFrame
frame "2" - BeginNavigation request to 'http://127.0.0.1:8000/loading/resources/redirect-methods-result.php', http method POST
frame "2" - DidStartNavigation
frame "2" - ReadyToCommitNavigation
frame "2" - didCommitLoadForFrame
frame "2" - didFinishDocumentLoadForFrame
frame "3" - BeginNavigation request to 'about:blank', http method GET
frame "3" - DidStartNavigation
frame "3" - ReadyToCommitNavigation
frame "3" - didCommitLoadForFrame
frame "3" - didFinishDocumentLoadForFrame
frame "3" - didHandleOnloadEventsForFrame
frame "3" - didFinishLoadForFrame
frame "3" - BeginNavigation request to 'http://127.0.0.1:8000/loading/resources/redirect-methods-form.html', http method GET
frame "3" - DidStartNavigation
frame "2" - didHandleOnloadEventsForFrame
frame "2" - didFinishLoadForFrame
frame "3" - ReadyToCommitNavigation
frame "3" - didCommitLoadForFrame
frame "3" - didFinishDocumentLoadForFrame
frame "3" - didHandleOnloadEventsForFrame
frame "3" - didFinishLoadForFrame
frame "3" - BeginNavigation request to 'http://127.0.0.1:8000/loading/resources/redirect-methods-result.php', http method POST
frame "3" - DidStartNavigation
frame "3" - ReadyToCommitNavigation
frame "3" - didCommitLoadForFrame
frame "3" - didFinishDocumentLoadForFrame
frame "3" - didHandleOnloadEventsForFrame
frame "3" - didFinishLoadForFrame
This test checks to see what HTTP method is used to fetch the final resource in the case where the first request results in a redirect.
301, 302, 303, and 307 http redirects are all tested.

301 redirect
302 redirect
303 redirect
307 redirect
